+++ src/extensions/score_metamodel/BUILD
@@ -56,6 +56,7 @@ py_library(
     # TODO: Figure out if all requirements are needed or if we can break it down a bit
     deps = all_requirements + [
         "@score_docs_as_code//src/helper_lib",
+        "@score_docs_as_code//src/extensions/score_metrics:score_metrics",
     ],
 )
