--- src/extensions/score_metamodel/traceability_metrics.py
@@ -1,258 +1,0 @@
-# *******************************************************************************
-# Copyright (c) 2026 Contributors to the Eclipse Foundation
-#
-# See the NOTICE file(s) distributed with this work for additional
-# information regarding copyright ownership.
-#
-# This program and the accompanying materials are made available under the
-# terms of the Apache License Version 2.0 which is available at
-# https://www.apache.org/licenses/LICENSE-2.0
-#
-# SPDX-License-Identifier: Apache-2.0
-# *******************************************************************************
-
-# ╓                                                          ╖
-# ║ Some portions generated by Github Copilot                ║
-# ╙                                                          ╜
-
-"""Shared traceability metric calculations for CI checks and dashboards."""
-
-from __future__ import annotations
-
-from collections.abc import Sequence
-from typing import Any
-
-
-def is_non_empty(value: Any) -> bool:
-    """Return True if value should be treated as present for traceability checks."""
-    if isinstance(value, str):
-        return bool(value.strip())
-    return bool(value)
-
-
-def parse_need_id_list(value: Any) -> list[str]:
-    """Normalize need-id lists encoded as CSV strings or string lists."""
-    if value is None:
-        return []
-    if isinstance(value, str):
-        return [item.strip() for item in value.split(",") if item.strip()]
-    if isinstance(value, list):
-        out: list[str] = []
-        for item in value:
-            if isinstance(item, str) and item.strip():
-                out.append(item.strip())
-        return out
-    return []
-
-
-def safe_percent(numerator: int, denominator: int) -> float:
-    """Return percentage in range [0, 100], treating empty denominator as 100%."""
-    if denominator == 0:
-        return 100.0
-    return (numerator / denominator) * 100.0
-
-
-def filter_requirements(
-    all_needs: Sequence[Any],
-    requirement_types: set[str],
-    include_not_implemented: bool,
-    include_external: bool = False,
-) -> list[Any]:
-    """Extract requirements by type, implementation state, and origin."""
-    requirements: list[dict[str, Any]] = []
-    for need in all_needs:
-        need_type = str(need.get("type", "")).strip()
-        if need_type not in requirement_types:
-            continue
-        if not include_external and need.get("is_external", False):
-            continue
-        if not include_not_implemented:
-            implemented = str(need.get("implemented", "")).upper().strip()
-            if implemented not in {"YES", "PARTIAL"}:
-                continue
-        requirements.append(need)
-    return requirements
-
-
-def calculate_requirement_metrics(
-    requirements: Sequence[Any],
-) -> dict[str, Any]:
-    """Calculate requirement traceability statistics for links and completeness."""
-    total = len(requirements)
-    with_code = sum(
-        1 for need in requirements if is_non_empty(need.get("source_code_link"))
-    )
-    with_test = sum(1 for need in requirements if is_non_empty(need.get("testlink")))
-    fully_linked = sum(
-        1
-        for need in requirements
-        if is_non_empty(need.get("source_code_link"))
-        and is_non_empty(need.get("testlink"))
-    )
-
-    missing_code_ids = [
-        str(need.get("id", ""))
-        for need in requirements
-        if not is_non_empty(need.get("source_code_link")) and need.get("id")
-    ]
-    missing_test_ids = [
-        str(need.get("id", ""))
-        for need in requirements
-        if not is_non_empty(need.get("testlink")) and need.get("id")
-    ]
-    not_fully_linked_ids = [
-        str(need.get("id", ""))
-        for need in requirements
-        if (
-            (
-                not is_non_empty(need.get("source_code_link"))
-                or not is_non_empty(need.get("testlink"))
-            )
-            and need.get("id")
-        )
-    ]
-
-    return {
-        "total": total,
-        "with_code_link": with_code,
-        "with_test_link": with_test,
-        "fully_linked": fully_linked,
-        "with_code_link_pct": safe_percent(with_code, total),
-        "with_test_link_pct": safe_percent(with_test, total),
-        "fully_linked_pct": safe_percent(fully_linked, total),
-        "missing_code_link_ids": sorted(missing_code_ids),
-        "missing_test_link_ids": sorted(missing_test_ids),
-        "not_fully_linked_ids": sorted(not_fully_linked_ids),
-    }
-
-
-def calculate_test_metrics(
-    all_needs: Sequence[Any],
-    requirement_ids: set[str],
-    filtered_test_types: set[str],
-) -> dict[str, Any]:
-    """Calculate testcase linkage and broken testcase-reference statistics."""
-    testcases = [
-        need for need in all_needs if str(need.get("type", "")).strip() == "testcase"
-    ]
-    if filtered_test_types:
-        testcases = [
-            need
-            for need in testcases
-            if str(need.get("test_type", need.get("TestType", ""))).strip()
-            in filtered_test_types
-        ]
-
-    tests_total = len(testcases)
-    tests_linked = 0
-    broken_references: list[dict[str, str]] = []
-
-    for test_need in testcases:
-        test_id = str(test_need.get("id", "<unknown_testcase>"))
-        partially = parse_need_id_list(
-            test_need.get("partially_verifies", test_need.get("PartiallyVerifies"))
-        )
-        fully = parse_need_id_list(
-            test_need.get("fully_verifies", test_need.get("FullyVerifies"))
-        )
-        refs = partially + fully
-        if refs:
-            tests_linked += 1
-        for ref in refs:
-            if ref not in requirement_ids:
-                broken_references.append({"testcase": test_id, "missing_need": ref})
-
-    return {
-        "total": tests_total,
-        "filtered_test_types": sorted(filtered_test_types),
-        "linked_to_requirements": tests_linked,
-        "linked_to_requirements_pct": safe_percent(tests_linked, tests_total),
-        "broken_references": broken_references,
-    }
-
-
-def calculate_process_requirement_metrics(
-    all_needs: Sequence[Any],
-    include_not_implemented: bool,
-    include_external: bool,
-) -> dict[str, Any]:
-    """Calculate process-requirement coverage via tool_req ``satisfies`` links."""
-    process_requirements = [
-        need
-        for need in all_needs
-        if str(need.get("type", "")).strip() in {"gd_req", "process_req"}
-        and (include_external or not need.get("is_external", False))
-    ]
-    process_requirement_ids = {
-        str(need.get("id", "")).strip()
-        for need in process_requirements
-        if need.get("id")
-    }
-
-    tool_requirements = filter_requirements(
-        all_needs,
-        requirement_types={"tool_req"},
-        include_not_implemented=include_not_implemented,
-        include_external=include_external,
-    )
-
-    linked_process_requirement_ids: set[str] = set()
-    for need in tool_requirements:
-        satisfies_ids = parse_need_id_list(need.get("satisfies", need.get("Satisfies")))
-        for ref_id in satisfies_ids:
-            if ref_id in process_requirement_ids:
-                linked_process_requirement_ids.add(ref_id)
-
-    total = len(process_requirement_ids)
-    linked_by_tool_requirements = len(linked_process_requirement_ids)
-    unlinked_ids = sorted(process_requirement_ids - linked_process_requirement_ids)
-
-    return {
-        "total": total,
-        "linked": linked_by_tool_requirements,
-        "linked_by_tool_requirements": linked_by_tool_requirements,
-        "linked_by_tool_requirements_pct": safe_percent(
-            linked_by_tool_requirements, total
-        ),
-        "unlinked_ids": unlinked_ids,
-    }
-
-
-def compute_traceability_summary(
-    all_needs: Sequence[Any],
-    requirement_types: set[str],
-    include_not_implemented: bool,
-    filtered_test_types: set[str],
-    include_external: bool = False,
-) -> dict[str, Any]:
-    """Return full CI/dashboard summary using one shared metric implementation."""
-    requirements = filter_requirements(
-        all_needs,
-        requirement_types=requirement_types,
-        include_not_implemented=include_not_implemented,
-        include_external=include_external,
-    )
-    requirement_ids = {
-        str(need.get("id", "")).strip() for need in requirements if need.get("id")
-    }
-
-    req_metrics = calculate_requirement_metrics(requirements)
-    test_metrics = calculate_test_metrics(
-        all_needs,
-        requirement_ids=requirement_ids,
-        filtered_test_types=filtered_test_types,
-    )
-    process_requirement_metrics = calculate_process_requirement_metrics(
-        all_needs,
-        include_not_implemented=include_not_implemented,
-        include_external=include_external,
-    )
-
-    return {
-        "requirement_types": sorted(requirement_types),
-        "include_not_implemented": include_not_implemented,
-        "include_external": include_external,
-        "requirements": req_metrics,
-        "tests": test_metrics,
-        "process_requirements": process_requirement_metrics,
-    }
