+++ src/extensions/score_metrics/traceability_metrics.py
@@ -1,0 +1,184 @@
+# *******************************************************************************
+# Copyright (c) 2026 Contributors to the Eclipse Foundation
+#
+# See the NOTICE file(s) distributed with this work for additional
+# information regarding copyright ownership.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Apache License Version 2.0 which is available at
+# https://www.apache.org/licenses/LICENSE-2.0
+#
+# SPDX-License-Identifier: Apache-2.0
+# *******************************************************************************
+
+# ╓                                                          ╖
+# ║ Some portions generated by Github Copilot                ║
+# ╙                                                          ╜
+
+"""Shared traceability metric calculations for CI checks and dashboards."""
+
+from __future__ import annotations
+
+from typing import Any
+from sphinx_needs.need_item import NeedItem
+from sphinx_needs.data import NeedsView
+
+
+def is_non_empty(value: Any) -> bool:
+    """Return True if value should be treated as present for traceability checks."""
+    if isinstance(value, str):
+        return bool(value.strip())
+    return bool(value)
+
+
+def safe_percent(numerator: int, denominator: int) -> float:
+    """Return percentage in range [0, 100], treating empty denominator as 100%."""
+    if denominator == 0:
+        return 100.0
+    return (numerator / denominator) * 100.0
+
+
+def calculate_requirement_metrics(
+    all_requirement_needs: list[NeedItem],
+) -> dict[str, Any]:
+    """Calculate requirement traceability statistics for links and completeness."""
+    total = len(all_requirement_needs)
+    reqs_with_code_link = 0
+    reqs_with_test_link = 0
+    reqs_fully_linked = 0
+    missing_code_links_need_ids: set[str] = set()
+    missing_test_links_need_ids: set[str] = set()
+    missing_both_links_need_ids: set[str] = set()
+
+    for need in all_requirement_needs:
+        # Sourcecode link check
+        if is_non_empty(need.get("source_code_link")):
+            reqs_with_code_link += 1
+        else:
+            missing_code_links_need_ids.add(need.get("id"))
+
+        # Testlink check
+        if is_non_empty(need.get("testlink")):
+            reqs_with_test_link += 1
+        else:
+            missing_test_links_need_ids.add(need.get("id"))
+        # Negative check (both missing)
+        if is_non_empty(need.get("testlink")) and is_non_empty(
+            need.get("source_code_link")
+        ):
+            reqs_fully_linked += 1
+
+    missing_both_links_need_ids = missing_code_links_need_ids.intersection(
+        missing_test_links_need_ids
+    )
+
+    return {
+        "total": total,
+        "with_code_link": reqs_with_code_link,
+        "with_test_link": reqs_with_test_link,
+        "fully_linked": reqs_fully_linked,
+        "with_code_link_pct": safe_percent(reqs_with_code_link, total),
+        "with_test_link_pct": safe_percent(reqs_with_test_link, total),
+        "fully_linked_pct": safe_percent(reqs_fully_linked, total),
+        "missing_code_link_ids": sorted(missing_code_links_need_ids),
+        "missing_test_link_ids": sorted(missing_test_links_need_ids),
+        "not_fully_linked_ids": sorted(missing_both_links_need_ids),
+    }
+
+
+def calculate_test_metrics(
+    all_needs: NeedsView,
+    filtered_test_types: set[str],
+) -> dict[str, Any]:
+    """Calculate testcase linkage and broken testcase-reference statistics."""
+    test_needs = all_needs.filter_types(["testcase"]).values()
+    tests_total = len(test_needs)
+    tests_linked = 0
+    broken_references: list[dict[str, str]] = []
+
+    for test_need in test_needs:
+        # This should never happen (that it has unknown_testcase id
+        # TODO: Remove this:
+        test_id = str(test_need.get("id", "<unknown_testcase>"))
+        partially: list[str] = test_need.get("partially_verifies")
+        fully: list[str] = test_need.get("fully_verifies")
+        refs = set(partially + fully)
+        if refs:
+            tests_linked += 1
+        for ref in refs:
+            if ref not in all_needs:
+                broken_references.append({"testcase": test_id, "missing_need": ref})
+
+    return {
+        "total": tests_total,
+        "filtered_test_types": sorted(filtered_test_types),
+        "linked_to_requirements": tests_linked,
+        "linked_to_requirements_pct": safe_percent(tests_linked, tests_total),
+        "broken_references": broken_references,
+    }
+
+
+def calculate_process_requirement_metrics(
+    all_needs: NeedsView,
+    include_external: bool,
+) -> dict[str, Any]:
+    """Calculate process-requirement coverage via tool_req ``satisfies`` links."""
+
+    process_requirements = all_needs.filter_types(["gd_req"]).filter_is_external(
+        include_external
+    )
+    process_requirement_ids = set(process_requirements.keys())
+    tool_requirements = all_needs.filter_types(["tool_req"]).filter_is_external(
+        include_external
+    )
+
+    linked_process_requirement_ids: set[str] = set()
+    for need in tool_requirements.values():
+        satisfies_ids = need.get("satisfies")
+        for ref_id in satisfies_ids:
+            if ref_id in process_requirement_ids:
+                linked_process_requirement_ids.add(ref_id)
+
+    total = len(process_requirement_ids)
+    linked_by_tool_requirements = len(linked_process_requirement_ids)
+    unlinked_ids = sorted(process_requirement_ids - linked_process_requirement_ids)
+
+    return {
+        "total": total,
+        "linked": linked_by_tool_requirements,
+        "linked_by_tool_requirements": linked_by_tool_requirements,
+        "linked_by_tool_requirements_pct": safe_percent(
+            linked_by_tool_requirements, total
+        ),
+        "unlinked_ids": unlinked_ids,
+    }
+
+
+def compute_traceability_summary(
+    all_needs: NeedsView,
+    filtered_test_types: set[str],
+    include_external: bool = False,
+) -> dict[str, Any]:
+    """Return full CI/dashboard summary using one shared metric implementation."""
+    requirements = all_needs
+    req_metrics = calculate_requirement_metrics(
+        list(requirements.filter_is_external(False).values())
+    )
+
+    test_metrics = calculate_test_metrics(
+        all_needs,
+        filtered_test_types=filtered_test_types,
+    )
+    process_requirement_metrics = calculate_process_requirement_metrics(
+        all_needs,
+        include_external=include_external,
+    )
+
+    return {
+        # Note: Unsure if we truly need this here. It seems to me to create too much spam
+        # "requirement_types": sorted(requirement_types),
+        "include_external": include_external,
+        "requirements": req_metrics,
+        "tests": test_metrics,
+        "process_requirements": process_requirement_metrics,
+    }
