+++ src/extensions/score_sphinx_bundle/__init__.py
@@ -31,6 +31,7 @@ score_extensions = [
     "sphinxcontrib.mermaid",
     "needs_config_writer",
     "score_sync_toml",
+    "score_metrics"
 ]
